# openapi/shipflow-api.yaml
openapi: 3.0.3
info:
  title: ShipFlow API
  version: 1.0.0
  description: |
    ShipFlow 跨境物流履约与费用结算系统第一版 API 契约。
    所有租户业务接口从 JWT 身份上下文获取 tenant_id，不接受调用方任意传入。
servers:
  - url: /api/v1
    description: ShipFlow API v1
...
security:
  - bearerAuth: []
paths:
...
  /platform/tenants/{tenantId}:
    get:
      tags: [Tenant]
      summary: 查询租户详情
      operationId: getTenant
      parameters: [ { $ref: '#/components/parameters/TenantId' } ]
      responses:
        '200': { description: 租户详情, headers: { X-Trace-Id: { $ref: '#/components/headers/TraceId' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessTenant' } } } }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Tenant]
      summary: 修改租户基础信息
      operationId: updateTenant
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateTenantRequest' } } }
      responses:
        '200': { description: 更新后租户, headers: { X-Trace-Id: { $ref: '#/components/headers/TraceId' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessTenant' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /logistics/channels:
    get:
      tags: [LogisticsMaster]
      summary: 租户分页查询公开渠道目录
      operationId: listAvailableLogisticsChannels
      parameters:
        - { name: channelCode, in: query, schema: { type: string, maxLength: 64 } }
        - { name: channelName, in: query, schema: { type: string, maxLength: 128 } }
        - { name: serviceCountry, in: query, schema: { type: string, pattern: '^[A-Z]{2}$' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ActiveStatus' } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: sortField, in: query, schema: { type: string, enum: [channelCode, channelName, providerName, status, effectiveFrom, updatedAt], default: updatedAt } }
        - { name: sortDirection, in: query, schema: { type: string, enum: [ASC, DESC], default: DESC } }
      responses:
        '200': { description: 公开渠道分页，不含价格阶梯、成本或供应商配置, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessPublicLogisticsChannelPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /logistics/channels/{channelId}:
    get:
      tags: [LogisticsMaster]
      summary: 租户查询公开渠道详情
      operationId: getAvailableLogisticsChannel
      parameters: [{ $ref: '#/components/parameters/ChannelId' }]
      responses:
        '200': { description: 公开渠道详情，不含价格阶梯、成本或供应商配置, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessPublicLogisticsChannel' } } } }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /logistics/channels/{channelId}/service-countries:
    get:
      tags: [LogisticsMaster]
      summary: 租户查询可用渠道服务国家
      operationId: getAvailableChannelServiceCountries
      description: 仅供具备 `scope:TENANT` 与 `logistics:read` 权限的租户用户查询当前启用渠道的服务国家代码。
      parameters: [{ $ref: '#/components/parameters/ChannelId' }]
      responses:
        '200': { description: ISO 3166-1 alpha-2 国家编码列表, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessServiceCountryList' } } } }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: 查询订单详情
      operationId: getOrder
      description: 仅查询 JWT tenant_id 对应租户的订单，需要 `scope:TENANT`。
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: 订单详情
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessShipmentOrder'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Orders]
      summary: 更新草稿订单的非计价信息
      operationId: updateDraftShipmentOrder
      description: |
        仅允许当前租户的 DRAFT 订单更新收件地址、寄件地址、商品明细和备注，需要 `scope:TENANT`。
        渠道、目的国家、申报重量尺寸、包裹数、价格规则版本、计费重量、金额、币种和报价快照均禁止修改；传入这些字段返回 422。
        version 是必填字段并用于订单乐观锁；版本过期或订单已不再是 DRAFT 时返回 409。
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShipmentOrderRequest'
      responses:
        '200': { description: 更新后的草稿订单, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessShipmentOrder' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/OrderRuleError' }
...
  /orders/{orderId}/price-confirmation:
    get:
      tags: [Orders]
      summary: 查询订单费用确认状态
      operationId: getPriceConfirmation
      description: |
        返回订单费用调整、申请状态和 UTC ISO-8601 时间字段。服务端按 tenant_id、store_id、角色和资源归属过滤。
        数据库存储时间统一为 UTC；前端展示和业务“今日”统计统一转换为 Asia/Shanghai。
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: 费用确认状态
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessPriceConfirmation'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/Forbidden'


        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Orders]
      summary: 确认复称后的新费用
      operationId: confirmPrice
      description: |
        财务人员最终确认费用，租户管理员可作为备用审批人；商家业务员只能提交申请，不能最终确认。
        客户端不能决定最终费用，服务端必须校验 feeAdjustmentId 属于当前订单和租户、类型为 INCREASE、尚未确认，
        expectedFee 等于 adjustment.after_amount，且订单当前状态为 PENDING_PRICE_CONFIRMATION。
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceConfirmationRequest'
      responses:
        '200':
          description: 新费用确认成功
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessPriceConfirmation'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/OrderRuleError'
  /orders/{orderId}/price-confirmation-requests:
    post:
      tags: [Orders]
      summary: 提交费用确认申请
      operationId: submitPriceConfirmationRequest
      description: 商家业务员只能在被授权店铺内提交申请；申请不改变订单状态，订单仍为 PENDING_PRICE_CONFIRMATION。
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceConfirmationRequest'
      responses:
        '200':
          description: 费用确认申请已提交
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessPriceConfirmation'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/OrderRuleError'
...
  /operations/summary:
    get:
      tags: [Operations]
      summary: 查询当前租户订单履约概览
      operationId: getOperationsSummary
      responses: { '200': { description: 各订单状态数量, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessOperationsSummary' } } } }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /operations/todos:
    get:
      tags: [Operations]
      summary: 查询当前租户跨域待办汇总
      operationId: getOperationsTodos
      responses: { '200': { description: 价格、对账、异常和索赔待办数量, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessOperationsTodos' } } } }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /operations/workbench:
    get:
      tags: [Operations]
      summary: 查询商家业务员运营概览工作台快照
      operationId: getOperationsWorkbench
      description: |
        只读概览快照。tenant_id 与 user_id 由 JWT 认证上下文取得，storeId 只能缩小后端重新计算的可见店铺范围。
        商家业务员仅能查看 ACTIVE 授权店铺，租户管理员及既有租户运营角色仅能查看本租户 ACTIVE 店铺。
        数据库时间按 UTC 存储和查询；TODAY/LAST_7_DAYS/LAST_30_DAYS 先按 Asia/Shanghai 计算自然日，再返回实际 UTC 窗口。
        refreshedAt、timeRange.from/to、订单和风险时间均为 UTC ISO-8601；金额使用 BigDecimal 两位小数语义，重量使用 kg。
      parameters:
        - name: timeRange
          in: query
          schema: { type: string, enum: [TODAY, LAST_7_DAYS, LAST_30_DAYS, CUSTOM], default: TODAY }
        - name: from
          in: query
          description: 仅 CUSTOM 使用，必须带 UTC offset 或 Z。
          schema: { $ref: '#/components/schemas/UtcDateTime' }
        - name: to
          in: query
          description: 仅 CUSTOM 使用，必须晚于 from，必须带 UTC offset 或 Z。
          schema: { $ref: '#/components/schemas/UtcDateTime' }
        - name: storeId
          in: query
          description: 可选店铺范围收窄条件，后端会再次校验租户、角色和 ACTIVE 授权归属。
          schema: { type: integer, format: int64, minimum: 1 }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: sortBy
          in: query
          schema: { type: string, enum: [updatedAt, createdAt], default: updatedAt }
        - name: sortDirection
          in: query
          schema: { type: string, enum: [ASC, DESC], default: DESC }
        - name: recentLimit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
        - name: riskLimit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        '200': { description: 与同一租户/店铺授权范围和时间窗口一致的工作台快照, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessOperationsWorkbench' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/OrderRuleError' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: REFRESH_TOKEN
      description: HttpOnly Refresh Token Cookie；客户端不可读取。
    providerHmac:
      type: apiKey
      in: header
      name: X-Provider-Signature
      description: Base64 编码的 HMAC-SHA256 签名，签名原文为 timestamp + "." + rawRequestBody。
  headers:
    TraceId:
      description: 与响应体 traceId 一致的链路追踪 ID。
      schema:
        type: string
    NoStore:
      description: 禁止缓存认证凭证相关响应。
      schema:
        type: string
        enum: [no-store]
        default: no-store
    WWWAuthenticate:
      description: RFC 6750 Bearer 认证挑战；具体错误详情不在 Header 中泄露敏感信息。
      schema:
        type: string
        example: Bearer
...
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
  responses:
    BadRequest:
      description: 请求参数或格式错误。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    AuthError:
      description: 未登录、Token 无效或已过期。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Forbidden:
      description: 权限不足或跨租户访问。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 资源不存在。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 幂等、乐观锁或订单状态冲突。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    OrderRuleError:
      description: 订单业务规则校验失败。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    Id:
      type: string
      pattern: '^[0-9]+$'
      example: '10001'
    Money:
      type: string
      pattern: '^[0-9]+\.[0-9]{2}$'
      example: '58.00'
      description: 非负金额，JSON 中使用字符串。
...
    UtcDateTime:
      type: string
      format: date-time
      example: '2026-08-02T10:30:00.000Z'
      description: UTC ISO-8601 时间。
...
    ApiSuccess:
      type: object
      required: [success, traceId, data]
      properties:
        success:
          type: boolean
          enum: [true]
        traceId:
          type: string
        message:
          type: string
          example: OK
        data: {}
...
    ApiError:
      type: object
      required: [success, traceId, error]
      properties:
        success:
          type: boolean
          enum: [false]
        traceId:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: ORDER-1001
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    Address:
      type: object
      required: [contactName, phone, countryCode, city, addressLine1, postalCode]
      properties:
        contactName:
          type: string
        companyName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        countryCode:
          $ref: '#/components/schemas/CountryCode'
        stateProvince:
          type: string
        city:
          type: string
        district:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        postalCode:
          type: string
...
    ShipmentOrderItem:
      type: object
      additionalProperties: false
      required: [sku, productName, quantity, unitPrice, currency]
      properties:
        sku: { type: string, minLength: 1 }
        productName: { type: string, minLength: 1 }
        quantity: { type: number, format: decimal, exclusiveMinimum: true }
        unitPrice: { $ref: '#/components/schemas/Money' }
        currency: { type: string, minLength: 1 }
        hsCode: { type: string }
        countryOfOrigin: { type: string }
    UpdateShipmentOrderRequest:
      type: object
      additionalProperties: false
      required: [version, senderAddress, receiverAddress, items]
      properties:
        version: { type: integer, format: int64, minimum: 0 }
        senderAddress: { $ref: '#/components/schemas/Address' }
        receiverAddress: { $ref: '#/components/schemas/Address' }
        items:
          type: array
          items: { $ref: '#/components/schemas/ShipmentOrderItem' }
        remark: { type: string }
        channelId: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        destinationCountry: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        declaredWeight: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        declaredLength: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        declaredWidth: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        declaredHeight: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        packageCount: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        priceRuleVersion: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        chargeableWeight: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        amount: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        currency: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
        quoteSnapshot: { nullable: true, description: 禁止修改；非空返回 ORDER-1007。 }
    ShipmentOrder:
      type: object
      required: [id, orderNo, quoteId, status, estimatedFee, currency, chargeableWeight, version, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        orderNo: { type: string }
        quoteId: { $ref: '#/components/schemas/Id' }
        status: { $ref: '#/components/schemas/OrderStatus' }
        estimatedFee: { $ref: '#/components/schemas/Money' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        chargeableWeight: { $ref: '#/components/schemas/Weight' }
        version:
          type: integer
          format: int64
          minimum: 0
          description: 当前订单乐观锁版本；草稿更新、提交或取消成功后返回递增后的最新值。
        createdAt: { $ref: '#/components/schemas/UtcDateTime' }
    ApiSuccessShipmentOrder:
      allOf:
        - $ref: '#/components/schemas/ApiSuccess'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ShipmentOrder' }
...
    PriceConfirmationView:
      type: object
      required: [id, orderNo, quoteId, storeId, currentStatus, estimatedFee, currentFee, currency, version, feeAdjustmentId, adjustmentType, beforeAmount, afterAmount, differenceAmount, confirmationStatus]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        orderNo: { type: string }
        quoteId: { $ref: '#/components/schemas/Id' }
        storeId: { $ref: '#/components/schemas/Id' }
        currentStatus: { $ref: '#/components/schemas/OrderStatus' }
        estimatedFee: { $ref: '#/components/schemas/Money' }
        currentFee: { $ref: '#/components/schemas/Money' }
        confirmedFee:
          allOf:
            - $ref: '#/components/schemas/Money'
          nullable: true
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        chargeableWeight: { $ref: '#/components/schemas/Weight' }
        version:
          type: integer
          format: int64
          minimum: 0
        feeAdjustmentId: { $ref: '#/components/schemas/Id' }
        adjustmentType: { type: string, enum: [INCREASE] }
        beforeAmount: { $ref: '#/components/schemas/Money' }
        afterAmount: { $ref: '#/components/schemas/Money' }
        differenceAmount: { $ref: '#/components/schemas/Money' }
        confirmationStatus:
          type: string
          enum: [PENDING_CONFIRMATION, REQUESTED, CONFIRMED]
        requestedAt:
          allOf:
            - $ref: '#/components/schemas/UtcDateTime'
          nullable: true
        confirmedAt:
          allOf:
            - $ref: '#/components/schemas/UtcDateTime'
          nullable: true
    ApiSuccessPriceConfirmation:
      allOf:
        - $ref: '#/components/schemas/ApiSuccess'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PriceConfirmationView'
    VersionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          format: int64
          minimum: 0
...
    PriceConfirmationRequest:
      type: object
      required: [feeAdjustmentId, expectedFee, version]
      properties:
        feeAdjustmentId:
          $ref: '#/components/schemas/Id'
        expectedFee:
          allOf:
            - $ref: '#/components/schemas/Money'
          description: 用于服务端校验的期望值，最终费用取 adjustment.after_amount。
        version:
          type: integer
          format: int64
          minimum: 0
...
    ActiveStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    TenantStatus:
      type: string
      enum: [PENDING, ACTIVE, DISABLED]
    Tenant:
      type: object
      required: [id, tenantCode, tenantName, status, version, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        tenantCode: { type: string, maxLength: 64 }
        tenantName: { type: string, maxLength: 128 }
        status: { $ref: '#/components/schemas/TenantStatus' }
        version: { type: integer, format: int64, minimum: 0 }
        createdAt: { $ref: '#/components/schemas/UtcDateTime' }
        updatedAt: { $ref: '#/components/schemas/UtcDateTime' }
    ApiSuccessTenant:
      allOf:
        - $ref: '#/components/schemas/ApiSuccess'
        - type: object
          properties: { data: { $ref: '#/components/schemas/Tenant' } }
...
    UpdateTenantRequest:
      allOf:
        - $ref: '#/components/schemas/VersionRequest'
        - type: object
          required: [tenantName]
          properties: { tenantName: { type: string, minLength: 1, maxLength: 128 } }
...
    ApiSuccessServiceCountryList:
      allOf:
        - $ref: '#/components/schemas/ApiSuccess'
        - type: object
          properties:
            data:
              type: array
              items: { type: string, pattern: '^[A-Z]{2}$' }
...
    PublicLogisticsChannel:
      type: object
      required: [id, providerName, channelCode, channelName, transportMode, serviceCountries, status, unavailableFields]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        providerName: { type: string, maxLength: 128 }
        channelCode: { type: string, maxLength: 64 }
        channelName: { type: string, maxLength: 128 }
        transportMode: { type: string, enum: [OCEAN, AIR, ROAD, RAIL, COURIER] }
        serviceCountries: { type: array, items: { type: string, pattern: '^[A-Z]{2}$' } }
        priceRuleVersion: { type: integer, minimum: 1, nullable: true, description: 当前生效公开规则版本号，不含价格阶梯 }
        effectiveFrom: { allOf: [{ $ref: '#/components/schemas/UtcDateTime' }], nullable: true }
        effectiveTo: { allOf: [{ $ref: '#/components/schemas/UtcDateTime' }], nullable: true }
        volumeDivisor: { type: number, format: decimal, nullable: true }
        status: { $ref: '#/components/schemas/ActiveStatus' }
        unavailableFields: { type: array, items: { type: string, enum: [supportedCargoAttributes, publicPriceDescription, priceRuleTiers, internalCost, supplierConfiguration] } }
    PublicLogisticsChannelPage:
      allOf: [{ $ref: '#/components/schemas/PageMeta' }, { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/PublicLogisticsChannel' } } } }]
    ApiSuccessPublicLogisticsChannel:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/PublicLogisticsChannel' } } }]
    ApiSuccessPublicLogisticsChannelPage:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/PublicLogisticsChannelPage' } } }]
...
    OperationsSummary:
      type: object
      properties: { draft: { type: integer, format: int64 }, pendingInbound: { type: integer, format: int64 }, inbound: { type: integer, format: int64 }, pendingPriceConfirmation: { type: integer, format: int64 }, readyForOutbound: { type: integer, format: int64 }, outbound: { type: integer, format: int64 }, inTransit: { type: integer, format: int64 }, delivered: { type: integer, format: int64 }, exception: { type: integer, format: int64 }, cancelled: { type: integer, format: int64 } }
...
    ApiSuccessOperationsSummary:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/OperationsSummary' } } }]
    ApiSuccessOperationsTodos:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/OperationsTodos' } } }]
...
    ApiSuccessOperationsWorkbench:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/OperationsWorkbench' } } }]
...